+++ app/components/my-test.js
@@ -1,0 +1,1 @@
+export { default } from '@jubileesoft/ember-jubileesoft-ui/components/my-test';