+++ config/environment.js
@@ -1,8 +1,5 @@
 'use strict';
 
 module.exports = function(/* environment, appConfig */) {
-  let ENV = {
-    podModulePrefix: 'ember-jubileesoft-ui/pods',
-  }
-  return ENV;
+  return {};
 };